#389 - Move configuration from packages
