+++ src/main/java/io/suricate/monitoring/utils/MonitorProxySelector.java
@@ -16,7 +16,7 @@
 
 package io.suricate.monitoring.utils;
 
-import io.suricate.monitoring.configuration.ProxyConfiguration;
+import io.suricate.monitoring.configuration.web.ProxyConfiguration;
 import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -42,10 +42,10 @@ public class MonitorProxySelector extends ProxySelector {
     public List<Proxy> select(URI uri) {
         Proxy ret = Proxy.NO_PROXY;
         ProxyConfiguration config = SpringContextHolder.getApplicationContext().getBean(ProxyConfiguration.class);
-        if(StringUtils.isNotBlank(config.getNoProxyDomains())) {
+        if (StringUtils.isNotBlank(config.getNoProxyDomains())) {
             try (Stream<String> stream = Arrays.stream(config.getNoProxyDomains().split(","))) {
-                if ( StringUtils.isNotBlank(config.getNoProxyDomains()) &&
-                     stream.noneMatch(h -> StringUtils.containsIgnoreCase(uri.getHost(), h)) ) {
+                if (StringUtils.isNotBlank(config.getNoProxyDomains()) &&
+                    stream.noneMatch(h -> StringUtils.containsIgnoreCase(uri.getHost(), h))) {
                     ret = new Proxy(Proxy.Type.HTTP, new InetSocketAddress(config.getHost(), Integer.valueOf(config.getPort())));
                 }
             }
